Version 3
I imported random numbers to complete the bandit branch. If you were to fight the bandits, you would have a 25% of winning and to hide, you would have a 50% of winningning. I didn't add this in the previous version, because adding random numbers posed a way different challenge. than the previous steps. With this implementation, all of the code is complete. Now comes corrections in the text, and cleaning it up.